+++ app/(main)/viewer/_views/Ready.tsx
@@ -2,7 +2,7 @@
 
 import MainButton from "@/app/_components/Main/MainButton";
 import { Container } from "./index.styled";
-import Config from "../_components/Config";
+import Config from "../../../_components/Viewer/Config";
 import { ViewersConfigType } from "@/lib/types";
 
 export default function Ready({
